requirements.txt
pinecone-client==3.0.2
python-dotenv==1.0.0
openai==1.12.0
python-telegram-bot==20.8
python-multipart==0.0.9
pydantic==2.6.1
loguru==0.7.2
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.22 